refactor: features/routing 리팩토링
