+++ src/app/route-guards/index.ts
@@ -1,1 +1,2 @@
 export { UserStateRouter } from './UserStateRouter'
+export { HomeRouter } from './HomeRouter'
