+++ src/views/home/ui/Layout.tsx
@@ -1,14 +1,7 @@
-import { HomeRouter } from '@/features/routing/HomeRouter'
-
 export function HomeLayout({
   children,
 }: Readonly<{
   children: React.ReactNode
 }>) {
-  return (
-    <>
-      <HomeRouter />
-      {children}
-    </>
-  )
+  return <>{children}</>
 }
